+++ Sources/SwiftUIKit/Images/Image+ImageRepresentable.swift
@@ -2,8 +2,8 @@
 //  Image+ImageRepresentable.swift
 //  SwiftUIKit
 //
-//  Created by Daniel Saidi on 2021-07-25.
-//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//  Created by Daniel Saidi on 2022-05-06.
+//  Copyright © 2022 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI